init commit of e2e testing with playwright and wp-env and refactor of Collectors.
## Testing
This commit will replace the old "behat" tests with playwright and wp-env.

## Refactor
The DataLayer and Collectors were refactored and are now required to be registered to the Service\DataCollectorRegistry which is being used by the DataLayer. Additionally, the Collectors are not require anymore to implement "enabled()" and "isAllowed". The DataLayer itself will contain the state of "enabled/disabled". Additionally, "Collector:data()" can now return "null", when the current view does not generate any data.

More changes:
- The AuthorDataCollector has been merged with the PostDataCollector.
- RendererProvider does not resolve dependencies lazy and not directly on registration.
- Http\Request has been removed.
- A new SearchDataCollector has been introduced.
- DataCollectorInterface has now "id()", "name()" and "description()" as new methods.