complete refactor of Plugin.
## removed dependencies
Following dependencies have been removed from the Plugin:

- inpsyde/assets
- chrico/wp-fields
- brain/wp-nonces

## Raise of required versions
PHP is now required in at least v8.1 and Node in v18.

## E2E Testing
E2E Testing with playwright has been added to the repository and is being executed in a GitHub Action through wp-env.

## Settings revamp
The Settings have been completely rewritten and are now managed with Gutenberg Components on the page. A new REST Endpoint has been introduced which is being used to get and update the settings.