3.0.0 - refactor (#30)
## Testing
This commit will replace the old "behat" tests with playwright and wp-env.

## Refactor DataLayer and Collectors
The DataLayer and Collectors were refactored and are now required to be registered to the Service\DataCollectorRegistry which is being used by the DataLayer. Additionally, the Collectors are not require anymore to implement "enabled()" and "isAllowed". The DataLayer itself will contain the state of "enabled/disabled" of Collectors. Additionally, "Collector:data()" can now return "null", when the current view does not generate any data.

More changes:
- The AuthorDataCollector has been merged with the PostDataCollector.
- RendererProvider does not resolve dependencies lazy and not directly on registration.
- Http\Request has been removed.
- A new SearchDataCollector has been introduced.
- DataCollectorInterface has now "id()", "name()" and "description()" as new methods.

## removed dependencies
Following dependencies have been removed from the Plugin:

- inpsyde/assets
- chrico/wp-fields
- brain/wp-nonces

## Raise of required versions
PHP is now required in at least v8.1 and Node in v18.

## E2E Testing
E2E Testing with playwright has been added to the repository and is being executed in a GitHub Action through wp-env.

## Settings revamp
The Settings have been completely rewritten and are now managed with Gutenberg Components on the page. A new REST Endpoint has been introduced which is being used to get and update the settings.


## Quality
* All PHP GA are now using min PHP 8.1
* All Node GA are now using Node 18.